// Front-End/src/app/models/module.concierge.vehicle.checklist.ts
export class ModuleConciergeVehicleChecklist {
    companyId: number | null = null;
    resaleId: number | null = null;
    id: number | null = null;

    checklist1Enabled: string | null = null;
    checklist1Desc: string = '';

    checklist2Enabled: string | null = null;
    checklist2Desc: string = '';

    checklist3Enabled: string | null = null;
    checklist3Desc: string = '';

    checklist4Enabled: string | null = null;
    checklist4Desc: string = '';

    checklist5Enabled: string | null = null;
    checklist5Desc: string = '';

    checklist6Enabled: string | null = null;
    checklist6Desc: string = '';

    checklist7Enabled: string | null = null;
    checklist7Desc: string = '';

    checklist8Enabled: string | null = null;
    checklist8Desc: string = '';

    checklist9Enabled: string | null = null;
    checklist9Desc: string = '';

    checklist10Enabled: string | null = null;
    checklist10Desc: string = '';

    checklist11Enabled: string | null = null;
    checklist11Desc: string = '';

    checklist12Enabled: string | null = null;
    checklist12Desc: string = '';

    checklist13Enabled: string | null = null;
    checklist13Desc: string = '';

    checklist14Enabled: string | null = null;
    checklist14Desc: string = '';

    checklist15Enabled: string | null = null;
    checklist15Desc: string = '';

    checklist16Enabled: string | null = null;
    checklist16Desc: string = '';

    checklist17Enabled: string | null = null;
    checklist17Desc: string = '';

    checklist18Enabled: string | null = null;
    checklist18Desc: string = '';

    checklist19Enabled: string | null = null;
    checklist19Desc: string = '';

    checklist20Enabled: string | null = null;
    checklist20Desc: string = '';
}